% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{addCountDataFromBam,DsATAC-method}
\alias{addCountDataFromBam,DsATAC-method}
\alias{addCountDataFromBam}
\title{addCountDataFromBam-methods}
\usage{
\S4method{addCountDataFromBam}{DsATAC}(.object, fns)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{fns}{a named vector of bam file locations. The names must correspond to sample identifiers in the object}
}
\value{
a new \code{\linkS4class{DsATAC}} object with read counts for each sample and region set
}
\description{
Add count data to DsATAC object based on bam files
}
\author{
Fabian Mueller
}
