% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{addCountDataFromGRL,DsATAC-method}
\alias{addCountDataFromGRL,DsATAC-method}
\alias{addCountDataFromGRL}
\title{addCountDataFromGRL-methods}
\usage{
\S4method{addCountDataFromGRL}{DsATAC}(.object, grl)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{grl}{NAMED \code{GRangesList} or NAMED list of \code{GRanges} objects. Names must correspond to sample ids in the object}
}
\value{
a new \code{\linkS4class{DsATAC}} object with read counts for each sample and region set
}
\description{
Add count data to DsATAC object based on a list of GRanges objects
}
\author{
Fabian Mueller
}
