% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsAcc-class.R
\docType{methods}
\name{addSampleAnnotCol,DsAcc-method}
\alias{addSampleAnnotCol,DsAcc-method}
\alias{addSampleAnnotCol}
\title{addSampleAnnotCol-methods}
\usage{
\S4method{addSampleAnnotCol}{DsAcc}(.object, name, vals)
}
\arguments{
\item{.object}{\code{\linkS4class{DsAcc}} object}

\item{name}{a name for the new column}

\item{vals}{vector of values}
}
\value{
a new \code{\linkS4class{DsAcc}} object with added sample annotation
}
\description{
add a sample annotation column to the sample annotation table
}
\author{
Fabian Mueller
}
