% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsASC-class.R
\name{ascAggregate}
\alias{ascAggregate}
\title{Aggregate ref/alt across a set of samples (columns)}
\usage{
ascAggregate(ref, alt, sampleIds)
}
\arguments{
\item{ref, alt}{count matrices [snp x sample].}

\item{sampleIds}{samples to sum over.}
}
\value{
list(ref, alt) row sums.
}
\description{
Aggregate ref/alt across a set of samples (columns)
}
