% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsASC-class.R
\name{ascFilterRegions}
\alias{ascFilterRegions}
\title{Drop ASC sites in blacklist / low-mappability regions}
\usage{
ascFilterRegions(snpIds, regionsGr, verbose = TRUE)
}
\arguments{
\item{snpIds}{character vector of "chr:pos_REF_ALT" ids (rownames of the
count matrices).}

\item{regionsGr}{GRanges of regions to remove (ENCODE blacklist and/or
low-mappability). Seqnames must be UCSC-style ("chr1", ...).}

\item{verbose}{log how many sites were removed (default TRUE).}
}
\value{
the subset of \code{snpIds} NOT overlapping \code{regionsGr}.
}
\description{
Repetitive and low-mappability regions are where WASP fails to remove reference
mapping bias AND where GATK makes false-heterozygous genotype calls (reads from
a paralog map in). Both push significant ASC toward the reference allele. Since
we lack the paper's imputed-genotype QC, removing these regions is the standard,
faithful mitigation (Calderon et al. also excluded blacklist regions in QC).
}
\author{
Irem B. GUNDUZ
}
