% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsASC-class.R
\name{ascStratum}
\alias{ascStratum}
\title{Stratify a pair of (cellType, stimulus) groups into the 4 sharing strata}
\usage{
ascStratum(cellA, condA, cellB, condB)
}
\description{
Stratify a pair of (cellType, stimulus) groups into the 4 sharing strata
}
