% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsASC-tfactivity.R
\name{ascTFFamily}
\alias{ascTFFamily}
\title{Assign a broad structural family to a (JASPAR) TF name}
\usage{
ascTFFamily(tf)
}
\arguments{
\item{tf}{character vector of TF names.}
}
\value{
character vector of family labels.
}
\description{
Coarse buckets used for colouring: "AP-1 / bZIP" (FOS/JUN/BATF/BACH/MAF/JDP/
ATF/NFE2, incl. dimers), "ZNF / GC" (ZNF/VEZF/RREB/GLIS/PLAGL/ZBTB/SP/KLF/EGR/
ZIC/WT1/CTCF), else "Other". Matching is case-insensitive and matches either
side of a dimer name (e.g. "FOSL1::JUND").
}
\author{
Irem B. GUNDUZ
}
