% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsASC-class.R
\name{calcASCStatistics}
\alias{calcASCStatistics}
\title{Calculate ASC Statistics (Binomial Test)}
\usage{
calcASCStatistics(dsObj, minCoverage = 10)
}
\arguments{
\item{dsObj}{A DsASC object (filtered).}

\item{minCoverage}{Minimum coverage to perform test (default 10).}
}
\value{
A data.table containing statistics for every SNP-Sample pair.
}
\description{
Performs a two-sided Binomial Test against p=0.5 for every SNP.
Adds P-values and FDR (Benjamini-Hochberg) to the result.
}
