% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChrAccR.R
\docType{package}
\name{ChrAccR-package}
\alias{ChrAccR}
\alias{ChrAccR-package}
\title{ChrAccR: Analyzing chromatin accessibility data in R.}
\description{
Tools for analyzing chromatin accessibility data in R. Currently supports ATAC-seq and NOMe-seq data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://epigenomeinformatics.github.io/ChrAccR}
  \item \url{https://github.com/EpigenomeInformatics/ChrAccR}
  \item Report bugs at \url{https://github.com/EpigenomeInformatics/ChrAccR/issues}
}

}
\author{
\strong{Maintainer}: Fabian Mueller \email{fabian.mueller@uni-saarland.de} (\href{https://orcid.org/0000-0001-5809-2321}{ORCID})

Authors:
\itemize{
  \item Irem B. G<U+00FC>nd<U+00FC>z \email{irembgunduz@gmail.com} (\href{https://orcid.org/0000-0003-2641-0916}{ORCID})
}

}
\keyword{internal}
