% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChrAccR.R
\docType{package}
\name{ChrAccR}
\alias{ChrAccR}
\title{ChrAccR: Analyzing chromatin accessibility data in R.}
\description{
Tools for analyzing chromatin accessibility data in R. Currently supports ATAC-seq and NOMe-seq data analysis.
}
