% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanMem}
\alias{cleanMem}
\title{cleanMem}
\usage{
cleanMem(iter.gc = 1L)
}
\arguments{
\item{iter.gc}{number of times to invoke the garbage collector}
}
\value{
nothing of particular interest
}
\description{
clean the system mory by invoking the garbage collector
}
\author{
Fabian Mueller
}

