% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_atac_differential.R
\docType{methods}
\name{createReport_differential,DsATAC-method}
\alias{createReport_differential,DsATAC-method}
\alias{createReport_differential}
\title{createReport_differential-methods}
\usage{
\S4method{createReport_differential}{DsATAC}(.object, reportDir)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{reportDir}{directory in which the report will be created}
}
\value{
(invisible) \code{muReportR::Report} object containing the report
}
\description{
Create a report summarizing differential accessibility analysis
}
\author{
Fabian Mueller
}
