% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_atac_filtering.R
\docType{methods}
\name{createReport_filtering,DsATAC-method}
\alias{createReport_filtering,DsATAC-method}
\alias{createReport_filtering}
\title{createReport_filtering-methods}
\usage{
\S4method{createReport_filtering}{DsATAC}(.object, reportDir, unfilteredObj, filterStats = NULL)
}
\arguments{
\item{.object}{filtered \code{\linkS4class{DsATAC}} object}

\item{reportDir}{directory in which the report will be created}

\item{unfilteredObj}{unfiltered \code{\linkS4class{DsATAC}} object}

\item{filterStats}{filtering statistics as output by \code{\link{run_atac_filtering}}}
}
\value{
(invisible) \code{muReportR::Report} object containing the report
}
\description{
Create a report summarizing steps and statistics
}
\author{
Fabian Mueller
}
