% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_atac_normalization.R
\docType{methods}
\name{createReport_normalization,DsATAC-method}
\alias{createReport_normalization,DsATAC-method}
\alias{createReport_normalization}
\title{createReport_normalization-methods}
\usage{
\S4method{createReport_normalization}{DsATAC}(.object, reportDir, unnormObj)
}
\arguments{
\item{.object}{normalized \code{\linkS4class{DsATAC}} object}

\item{reportDir}{directory in which the report will be created}

\item{unnormObj}{unnormalized \code{\linkS4class{DsATAC}} object}
}
\value{
(invisible) \code{muReportR::Report} object containing the report
}
\description{
Create a report summarizing normalization
}
\author{
Fabian Mueller
}
