% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_atac_summary.R
\docType{methods}
\name{createReport_summary,DsATAC-method}
\alias{createReport_summary,DsATAC-method}
\alias{createReport_summary}
\title{createReport_summary-methods}
\usage{
\S4method{createReport_summary}{DsATAC}(.object, reportDir)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{reportDir}{directory in which the report will be created}
}
\value{
(invisible) \code{muReportR::Report} object containing the report
}
\description{
Create a report summarizing an accessibility dataset
}
\examples{
\dontrun{
dsa <- ChrAccRex::loadExample("dsAtac_ia_example")
reportDir <- file.path(".", "ChrAccR_reports")
createReport_summary(dsa, reportDir)
}
}
\author{
Fabian Mueller
}
