% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_asc.R
\name{DsASC.gatk}
\alias{DsASC.gatk}
\title{Import Allele-Specific Chromatin Data (GATK VCFs + BAM Pileup)}
\usage{
DsASC.gatk(sampleAnnot, vcfDir, genome, diskDump = FALSE)
}
\arguments{
\item{sampleAnnot}{Data frame with sample annotation. Must contain 'sampleId', 'bamFilename', 'donor'.}

\item{vcfDir}{Directory containing donor VCFs named \code{\{donor\}_heterozygous.vcf.gz}.}

\item{genome}{Character string containing genome assembly (e.g. "hg38").}

\item{diskDump}{Logical. If TRUE, matrices are realized as HDF5 arrays.}
}
\value{
A \code{\linkS4class{DsASC}} object.
}
\description{
This function imports ASC data by counting reads at heterozygous SNP sites
defined in donor-specific VCF files. It uses \code{Rsamtools::pileup} to
quantify reference and alternative alleles.
}
