% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsASC-class.R
\name{DsASC}
\alias{DsASC}
\title{Create a new DsASC object}
\usage{
DsASC(sampleAnnot, genome, diskDump = FALSE)
}
\arguments{
\item{sampleAnnot}{data.frame with sample annotation.}

\item{genome}{Character string containing genome assembly (e.g. "hg38").}

\item{diskDump}{Logical. If TRUE, matrices are realized as HDF5 arrays.}
}
\value{
A \code{\linkS4class{DsASC}} object.
}
\description{
Create a new DsASC object
}
\author{
Irem B. GUNDUZ
}
