% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_atac.R
\name{DsATAC.snakeATAC}
\alias{DsATAC.snakeATAC}
\title{DsATAC.bw}
\usage{
DsATAC.snakeATAC(sampleAnnot, filePrefixCol, genome, dataDir,
  regionSets = NULL, sampleIdCol = filePrefixCol, type = "bam")
}
\arguments{
\item{sampleAnnot}{data.frame specifying the sample annotation table}

\item{filePrefixCol}{column name specifying the file prefix for each sample in the sample annotation table}

\item{genome}{genome assembly}

\item{dataDir}{directory where the files are located}

\item{regionSets}{a list of GRanges objects which contain region sets over which count data will be aggregated}

\item{sampleIdCol}{column name or index in the sample annotation table containing unique sample identifiers}

\item{type}{input data type. Currently only "bam" is supported}
}
\value{
\code{\linkS4class{DsATAC}} object
}
\description{
Create a DsATAC dataset from multiple input files in bigwig format
}
\author{
Fabian Mueller
}
