% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATACsc-class.R
\docType{class}
\name{DsATACsc-class}
\alias{DsATACsc-class}
\title{DsATACsc}
\description{
A class for storing single-cell ATAC-seq accessibility data
inherits from \code{\linkS4class{DsATAC}}. Provides a few additional methods
but is otherwise identical to \code{\linkS4class{DsATAC}}.
}
\author{
Fabian Mueller
}
