% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_atac_archr.R
\name{DsATACsc.archr}
\alias{DsATACsc.archr}
\title{DsATACsc.archr}
\usage{
DsATACsc.archr(
  ap,
  keepInsertionInfo = FALSE,
  diskDump.fragments = keepInsertionInfo
)
}
\arguments{
\item{ap}{\code{ArchR} project object}

\item{keepInsertionInfo}{flag indicating whether to maintain the insertion information in the resulting object.}

\item{diskDump.fragments}{Keep fragment coordinates stored on disk rather than in main memory. This saves memory, but increases runtime and I/O.}
}
\value{
\code{\linkS4class{DsATACsc}} object
}
\description{
Create a DsATACsc dataset from an \code{ArchR} project
}
\author{
Fabian Mueller
}
