% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsNOMe-class.R
\docType{class}
\name{DsNOMe-class}
\alias{DsNOMe-class}
\title{DsNOMe}
\description{
A class for storing NOMe accessibility data
}
\section{Slots}{

\describe{
  \item{\code{meth}}{
		List of GC methylation for sites and summarized regions.
  }
  \item{\code{covg}}{
		List of GC read coverage for sites and summarized regions.
  }
}
}

\author{
Fabian Mueller
}
