% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTools_nome.R
\name{DsNOMe.bisSNP}
\alias{DsNOMe.bisSNP}
\title{DsNOMe.bisSNP}
\usage{
DsNOMe.bisSNP(inputFns, sampleAnnot, genome,
  sampleIds = rownames(sampleAnnot))
}
\arguments{
\item{inputFns}{a NAMED vector of input file names}

\item{sampleAnnot}{data.frame spcifying the sample annotation table}

\item{genome}{genome assembly}

\item{sampleIds}{character vector of sample names}
}
\value{
\code{\linkS4class{Job}} object
}
\description{
Create a DsNOMe dataset from multiple input files in bisSNP output format
}
\author{
Fabian Mueller
}
