% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsASC-class.R
\name{estimateSharedImbalance}
\alias{estimateSharedImbalance}
\title{Estimate the Proportion of Shared ASC Imbalance Effects}
\usage{
estimateSharedImbalance(dsObjA, dsObjB, fdrCutoff = 0.01)
}
\arguments{
\item{dsObjA}{DsASC object for Cell Type A.}

\item{dsObjB}{DsASC object for Cell Type B.}

\item{fdrCutoff}{Numeric. FDR threshold for defining significance in Cell Type A (Default 0.01).}
}
\value{
A list with the ashR model output and the estimated proportion of shared effects.
}
\description{
Estimates the proportion of ASC sites (significant in Cell Type A) that still
show a non-zero effect in Cell Type B, using the Beta-Binomial posterior and ashR.
}
