% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_filehandling.R
\name{fastqDirToTable}
\alias{fastqDirToTable}
\title{fastqDirToTable}
\usage{
fastqDirToTable(fqDir, tabFn = NULL)
}
\arguments{
\item{fqDir}{string specifying a directory with fastq files}

\item{tabFn}{filename specifying the where the table should be written to. If NULL (default), the table will just be returned
as data frame}
}
\value{
data frame of parsed annotation
}
\description{
scan a directory containing fastq files and create a sample annotation table from the file names
}
\author{
Fabian Mueller
}
