% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{filterByGRanges,DsATAC-method}
\alias{filterByGRanges,DsATAC-method}
\alias{filterByGRanges}
\title{filterByGRanges-methods}
\usage{
\S4method{filterByGRanges}{DsATAC}(.object, gr, method = "black")
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{gr}{\code{GRanges} object used for filtering}

\item{method}{character string specifying the method. Can be \code{"white"} to treat \code{gr} as a whitelist
(i.e. only regions and fragments overlapping with it are retained) or \code{"black"} (default) to treat it as a blacklist.}
}
\value{
a new, filtered \code{\linkS4class{DsATAC}} object
}
\description{
Filter out regions based on a GRanges object
}
\author{
Fabian Mueller
}
