% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATACsc-class.R
\docType{methods}
\name{filterCellsTssEnrichment,DsATACsc-method}
\alias{filterCellsTssEnrichment,DsATACsc-method}
\alias{filterCellsTssEnrichment}
\alias{filterCellsTssEnrichment,DsATAC-method}
\title{filterCellsTssEnrichment-methods}
\usage{
\S4method{filterCellsTssEnrichment}{DsATACsc}(.object, cutoff = 6)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{cutoff}{TSS enrichment cutoff to filter cells}
}
\value{
modified \code{\linkS4class{DsATAC}} object without filtered cells
}
\description{
Filter out cells with low TSS enrichment
}
\author{
Fabian Mueller
}
