% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{filterChroms,DsATAC-method}
\alias{filterChroms,DsATAC-method}
\alias{filterChroms}
\title{filterChroms-methods}
\usage{
\S4method{filterChroms}{DsATAC}(.object, exclChrom = c("chrX", "chrY", "chrM"))
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{exclChrom}{vector of chromosome names to filter out}
}
\value{
a new \code{\linkS4class{DsATAC}} object filtered for chromosomes
}
\description{
Filter out regions based on chromosome list
}
\author{
Fabian Mueller
}
