% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsASC-class.R
\name{filterForRecurrence}
\alias{filterForRecurrence}
\title{Filter for Recurrent Allele-Specific Chromatin Events}
\usage{
filterForRecurrence(stats_dt, minDonors = 3, fdrCutoff = 0.05)
}
\arguments{
\item{stats_dt}{A data.table containing statistics (output of calcASCStatistics).}

\item{minDonors}{Integer. Minimum number of independent donors required (default 3).}

\item{fdrCutoff}{Numeric. FDR threshold to define significance (default 0.05).}
}
\value{
A data.table containing only the recurrent ASC events.
}
\description{
Identifies SNPs that are significantly imbalanced (ASC) in a minimum number
of independent donors.
}
