% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findGroupMarkers}
\alias{findGroupMarkers}
\title{findGroupMarkers}
\usage{
findGroupMarkers(
  X,
  grouping,
  bgRatio = 1,
  matchFrac = 1,
  testMethod = "wilcoxon"
)
}
\arguments{
\item{X}{matrix with values to be differentially tested for each group (features X samples)}

\item{grouping}{Group assignments. Ideally a factor vector. The length must match the number of samples (columns of X)}

\item{bgRatio}{For each group the number of samples that will be drawn as background set. This should be a factor, i.e. a value of 2 means that the background set for each group will be twice the size of the foreground set}

\item{matchFrac}{fraction of samples from the background set that are drawn from the pool of nearest neighbors of the foreground. This way the degree of matching between background and foreground can be adjusted.}

\item{testMethod}{test to be used. Currently valid options are \code{"wilcoxon"} and \code{"ttest"}}
}
\value{
list containing test statistics for each group and details on the foreground/background samples drawn for each group
}
\description{
Detect group markers by testing distribution of values corresponding to a group (foreground) to the corresponding values from a background set.
The background set can be matched by finding nearest neighbors to the foreground.
}
\author{
Fabian Mueller
}
