% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findNearestGeneForGr}
\alias{findNearestGeneForGr}
\title{findNearestGeneForGr}
\usage{
findNearestGeneForGr(gr, geneGr = NULL, maxDist = 1e+05)
}
\arguments{
\item{gr}{\code{GRanges} object}

\item{geneGr}{gene annotation from which to pull the annotation. Can be \code{NULL} for
automatic retrieval of annotation. Must be named or have a gene name column
in the metadata}

\item{maxDist}{maximum distance for matching to nearest gene}
}
\value{
\code{data.frame} containing information on the nearest gene for each element in \code{gr}
}
\description{
get gene annotation for a \code{GRanges} object by linking to the nearest gene
}
