% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsASC-class.R
\name{getAccessibility,DsASC-method}
\alias{getAccessibility,DsASC-method}
\title{Get the peak-level accessibility matrix}
\usage{
\S4method{getAccessibility}{DsASC}(object, asMatrix = TRUE, ...)
}
\arguments{
\item{object}{A \code{\linkS4class{DsASC}} object.}

\item{asMatrix}{Logical. Realize HDF5-backed arrays into a dense matrix.}
}
\value{
A matrix of peak accessibility counts.
}
\description{
Returns the total (allele-agnostic) ATAC-seq count matrix [peak x sample].
This is the chromatin accessibility signal used for differential
accessibility and for filtering SNPs to open regions.
}
\author{
Irem B. GUNDUZ
}
