% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{getConfigElement}
\alias{getConfigElement}
\title{getConfigElement}
\usage{
getConfigElement(name)
}
\arguments{
\item{name}{name of the config item}
}
\value{
the value of the config item
}
\description{
Get the value for a configuration item
}
\author{
Fabian Mueller
}

