% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsNOMe-class.R
\docType{methods}
\name{getCoord,DsNOMe-method}
\alias{getCoord,DsNOMe-method}
\alias{getCoord}
\title{getCoord-methods}
\usage{
\S4method{getCoord}{DsNOMe}(.object, type = "sites")
}
\arguments{
\item{.object}{\code{\linkS4class{DsNOMe}} object}

\item{type}{character string specifying the rgion type or \code{"sites"} (default)}

\item{asMatrix}{return a matrix instead of a \code{data.table}}
}
\value{
\code{GRanges} object containing coordinates for covered
        sites/regions
}
\description{
Return coordinates of sites/regions in a dataset
}
\author{
Fabian Mueller
}
