% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{getCoverage,DsATAC-method}
\alias{getCoverage,DsATAC-method}
\alias{getCoverage}
\title{getCoverage-methods}
\usage{
\S4method{getCoverage}{DsATAC}(.object, samples = getSamples(.object))
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{samples}{sample identifiers}
}
\value{
\code{list} of \code{Rle} objects of sample coverage tracks
}
\description{
Return a list of genome-wide coverage from insertion sites
}
\author{
Fabian Mueller
}
