% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATACsc-class.R
\docType{methods}
\name{getDiffAcc,DsATACsc-method}
\alias{getDiffAcc,DsATACsc-method}
\alias{getDiffAcc}
\alias{getDiffAcc,DsATAC-method}
\title{getDiffAcc-methods}
\usage{
\S4method{getDiffAcc}{DsATACsc}(
  .object,
  regionType,
  comparisonCol,
  grp1Name = NULL,
  grp2Name = NULL,
  nCellsPerBulk = 100,
  nBulkPerGroup = 20,
  method = "DESeq2"
)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATACsc}} object}

\item{regionType}{character string specifying the region type}

\item{comparisonCol}{column name in the cell annotation table to base the comparison on. Alternatively, a vector with one element for each
cell in the dataset that can be coerced to a factor}

\item{grp1Name}{name of the first group in the comparison. if not specified, it will be taken as the first factor level specified in the 
cell annotation (see \code{'comparisonCol'}).}

\item{grp2Name}{name of the second group (reference) in the comparison. if not specified, it will be taken as the first factor level specified in the 
cell annotation (see \code{'comparisonCol'}).}

\item{nCellsPerBulk}{number of cells to sample to create each pseudo-bulk sample}

\item{nBulkPerGroup}{number of pseudo-bulk samples to create for each group}

\item{method}{Method for determining differential accessibility. Currently only \code{'DESeq2'} is supported}
}
\value{
a \code{data.frame} containing differential accessibility information
}
\description{
Compute differential accessibility for single-cell datasets by randomly drawing cells from each group and aggregating them into pseudo-bulk samples
which are then compared using bulk differential methods
}
\author{
Fabian Mueller
}
