% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{getFragmentGr,DsATAC-method}
\alias{getFragmentGr,DsATAC-method}
\alias{getFragmentGr}
\title{getFragmentGr-methods}
\usage{
\S4method{getFragmentGr}{DsATAC}(.object, sampleId)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{sampleId}{sample identifier}
}
\value{
\code{GRanges} object containing fragment data
}
\description{
Return a \code{GRanges} object of fragment data for a given sample
}
\author{
Fabian Mueller
}
