% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{getFragmentGrl,DsATAC-method}
\alias{getFragmentGrl,DsATAC-method}
\alias{getFragmentGrl}
\title{getFragmentGrl-methods}
\usage{
\S4method{getFragmentGrl}{DsATAC}(.object, sampleIds,
  asGRangesList = FALSE)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{sampleIds}{sample identifiers}

\item{asGRangesList}{should a \code{GRangesList} object be returned instead of a regular \code{list}}
}
\value{
A named list of \code{GRanges} objects containing fragment data
}
\description{
Return a list of \code{GRanges} objects of fragment data for a given set of samples
}
\author{
Fabian Mueller
}
