% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{getFragmentNum,DsATAC-method}
\alias{getFragmentNum,DsATAC-method}
\alias{getFragmentNum}
\title{getFragmentNum-methods}
\usage{
\S4method{getFragmentNum}{DsATAC}(.object, sampleIds = getSamples(.object))
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{sampleIds}{sample identifiers}
}
\value{
a vector of fragment counts per sample
}
\description{
Return the number of fragments in the \code{\linkS4class{DsATAC}} object
}
\author{
Fabian Mueller
}
