% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsAcc-class.R
\docType{methods}
\name{getGenome,DsAcc-method}
\alias{getGenome,DsAcc-method}
\alias{getGenome}
\title{getGenome-methods}
\usage{
\S4method{getGenome}{DsAcc}(.object)
}
\arguments{
\item{.object}{\code{\linkS4class{DsAcc}} object}
}
\value{
Character string containing the genome assembly
}
\description{
Return the genome assembly
}
\author{
Fabian Mueller
}
