% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGroupsFromTable}
\alias{getGroupsFromTable}
\title{getGroupsFromTable}
\usage{
getGroupsFromTable(tt, cols = NULL, minGrpSize = 2, maxGrpCount = nrow(tt) - 1)
}
\arguments{
\item{tt}{table to retrieve groupings for}

\item{cols}{(Optional) predefined column names (in the form of a \code{character} vector) or indices (an
\code{integer} vector) to consider. All other columns in the annotation table will be ignored.}

\item{minGrpSize}{Minimum number of items required to form a group in comparison}

\item{maxGrpCount}{Maximum number of groups to be considered}
}
\value{
List of groupings. Each element corresponds to a categorical column in the table and contains the row indices for each
        category
}
\description{
Retrieve groupings given a table containing some categorical columns
}
\author{
Fabian Mueller
}
