% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{getInsertionKmerFreq,DsATAC-method}
\alias{getInsertionKmerFreq,DsATAC-method}
\alias{getInsertionKmerFreq}
\title{getInsertionKmerFreq-methods}
\usage{
\S4method{getInsertionKmerFreq}{DsATAC}(.object,
  samples = getSamples(.object), k = 6, normGenome = FALSE)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{samples}{sample identifiers}

\item{k}{length of the kmer}

\item{normGenome}{should the result be normalized by genome-wide kmer frequencies}
}
\value{
a \code{matrix} containing kmer frequencies (one row for each kmer and one column for each sample in the dataset)
}
\description{
compute kmer frequencies at insertion sites for each sample
}
\author{
Fabian Mueller
}
