% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{getInsertionSites,DsATAC-method}
\alias{getInsertionSites,DsATAC-method}
\alias{getInsertionSites}
\title{getInsertionSites-methods}
\usage{
\S4method{getInsertionSites}{DsATAC}(.object,
  samples = getSamples(.object))
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{samples}{sample identifiers}
}
\value{
\code{GRangesList} containing Tn5 cut sites for each sample
}
\description{
Return a list of insertion sites (Tn5 cut sites) for each sample
}
\author{
Fabian Mueller
}
