% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_motifs.R
\name{getJasparAnnot}
\alias{getJasparAnnot}
\title{getJasparAnnot}
\usage{
getJasparAnnot(ss, type = "humantfs")
}
\arguments{
\item{ss}{character vector or JASPAR identifiers}

\item{type}{annotation type. Currently only \code{"humantfs"} (pulls info from humantfs.ccbr.utoronto.ca) is supported}
}
\value{
list of data frames of TF annotation (a motif can have multiple annotated TFs)
}
\description{
retrieve motif annotation data
}
\author{
Fabian Mueller
}
