% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_motifs.R
\name{getJasparSymbols}
\alias{getJasparSymbols}
\title{getJasparSymbols}
\usage{
getJasparSymbols(ss)
}
\arguments{
\item{ss}{character vector or JASPAR identifiers}
}
\value{
a list of TF names (symbols) for each identifier
}
\description{
Retrieve the TF names (symbols) from a JASPAR identifier
}
