% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_motifs.R
\name{getMotifDistMat.jaspar}
\alias{getMotifDistMat.jaspar}
\title{getMotifDistMat.jaspar}
\usage{
getMotifDistMat.jaspar(motifIds = NULL, scoreCol = "Ncor")
}
\arguments{
\item{motifIds}{string vector of motif ids whose dissimilarities are retrieved}

\item{scoreCol}{namew of the annotation column in the JASPAR annotation that contains the motif similarity}
}
\value{
a matrix of motif DISsimilarities
}
\description{
Retrieve motif a comparison table from JASPAR annotation website and construct a dissimilarity matrix for given
motif IDs
}
