% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_motifs.R
\name{getMotifDistMat}
\alias{getMotifDistMat}
\title{getMotifDistMat}
\usage{
getMotifDistMat(assembly = "hg38", mmObj = NULL, method = "jaspar")
}
\arguments{
\item{assembly}{genome assembly for which the motifs dissimilarity should be retrieved. Only the species information
of the assembly is really relevant}

\item{mmObj}{optional motifmatchr object as returned by \code{ChrAccR::prepareMotifmatchr}}

\item{method}{method of dissimilarity quantification. Currently only \code{'jaspar'} (retrieve motif similarities from the annotation of the JASPAR website) is supported.}
}
\value{
a matrix of motif DISsimilarities (\code{dist} object)
}
\description{
Retrieve motif dissimilarity/distance matrix for TF motifs
}
\author{
Fabian Mueller
}
