% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{getMotifEnrichment,DsATAC-method}
\alias{getMotifEnrichment,DsATAC-method}
\alias{getMotifEnrichment}
\title{getMotifEnrichment-methods}
\usage{
\S4method{getMotifEnrichment}{DsATAC}(.object, type, idx, motifs = "jaspar")
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{type}{character string specifying the region type}

\item{idx}{logical vector or indices of the same length as \code{length(getCoord(.object))}
specifies the query set}

\item{motifs}{either a character string (currently only "jaspar" is supported) or an object containing PWMs
that can be used by \code{motifmatchr::matchMotifs} (such as an \code{PFMatrixList} object)}
}
\value{
a \code{data.frame} summarizing Fisher's Exact Test enrichment statistics for each motif
}
\description{
Perform enrichment analysis for (TF) motifs of a query set of regions.
Fisher's Exact Test is employed to test the association of motif present in the
query set against the background of all regions of that type covered in the object
}
\author{
Fabian Mueller
}
