% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_motifs.R
\name{getMotifOccurrences}
\alias{getMotifOccurrences}
\title{getMotifOccurrences}
\usage{
getMotifOccurrences(motifNames = NULL, motifDb = "jaspar",
  genome = "hg38")
}
\arguments{
\item{motifNames}{character vector of motif names}

\item{motifDb}{either a character string (currently only "jaspar" and sets contained in \code{chromVARmotifs} ("homer", "encode", "cisbp") are supported) or an object containing PWMs
that can be used by \code{motifmatchr::matchMotifs} (such as an \code{PFMatrixList} or \code{PWMatrixList} object)}

\item{genome}{character string specifying genome assembly}
}
\value{
a \code{GenomicRangesList} containing motif occurrences
}
\description{
Find occurrences of motifs in a given genome
}
\author{
Fabian Mueller
}
