% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_atac.R
\name{getNonOverlappingByScore}
\alias{getNonOverlappingByScore}
\title{getNonOverlappingByScore}
\usage{
getNonOverlappingByScore(gr, scoreCol = "score")
}
\arguments{
\item{gr}{\code{GRanges} object}

\item{scoreCol}{name of the column to be used as scor in the \code{elementMetadata} of the \code{gr} object}
}
\value{
\code{GRanges} object containing non-overlapping regions
}
\description{
Retrieve the set of non-verlapping regions by iteratively picking the region with maximum score for
each set of consecutively overlapping regions
}
\author{
Fabian Mueller
}
