% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsAcc-class.R
\docType{methods}
\name{getNRegions,DsAcc-method}
\alias{getNRegions,DsAcc-method}
\alias{getNRegions}
\title{getNRegions-methods}
\usage{
\S4method{getNRegions}{DsAcc}(.object, type = "sites")
}
\arguments{
\item{.object}{\code{\linkS4class{DsAcc}} object}

\item{type}{character string specifying the rgion type or \code{"sites"} (default)}
}
\value{
the number of regions of that type
}
\description{
Return the number of regions of a given type
}
\author{
Fabian Mueller
}
