% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsNOMe-class.R
\docType{methods}
\name{getNRegions,DsNOMe-method}
\alias{getNRegions,DsNOMe-method}
\alias{getNRegions}
\title{getNRegions-methods}
\usage{
\S4method{getNRegions}{DsNOMe}(.object, type = "sites")
}
\arguments{
\item{.object}{\code{\linkS4class{DsNOMe}} object}

\item{type}{character string specifying the rgion type or \code{"sites"} (default)}
}
\value{
the number of regions of that type
}
\description{
Return the number of regions of a given type
}
\author{
Fabian Mueller
}
