% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_atac.R
\name{getPeakSet.snakeATAC}
\alias{getPeakSet.snakeATAC}
\title{getPeakSet.snakeATAC}
\usage{
getPeakSet.snakeATAC(sampleAnnot, filePrefixCol, genome, dataDir,
  sampleIdCol = filePrefixCol, type = "summits_no_fw", unifWidth = 500L)
}
\arguments{
\item{sampleAnnot}{data.frame specifying the sample annotation table}

\item{filePrefixCol}{column name specifying the file prefix for each sample in the sample annotation table}

\item{genome}{genome assembly}

\item{dataDir}{directory where the files are located}

\item{sampleIdCol}{column name or index in the sample annotation table containing unique sample identifiers}

\item{type}{input data type. Currently only "summits_no_fw" (non-overlapping, fixed-width peaks deduced from summits)}

\item{unifWidth}{width of the peaks if the results have uniform peak lengths}
}
\value{
\code{GRanges} object containing consensus peak set
}
\description{
Retrieve a consensus set of ATAC peaks from the snakeATAC pipline run
}
\author{
Fabian Mueller
}
