% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsNOMe-class.R
\docType{methods}
\name{getRegionMapping,DsNOMe-method}
\alias{getRegionMapping}
\alias{getRegionMapping,DsNOMe-method}
\title{getRegionMapping-methods}
\usage{
\S4method{getRegionMapping}{DsNOMe}(.object, type)
}
\arguments{
\item{.object}{\code{\linkS4class{DsNOMe}} object}

\item{type}{character string specifying a name for the region type}
}
\value{
list containing vectors of indices of GCs for each region of the
        specified type
}
\description{
Retrieve a mapping from regions to GC indices in the dataset
}
\author{
Fabian Mueller
}

