% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsAcc-class.R
\docType{methods}
\name{getRegionTypes,DsAcc-method}
\alias{getRegionTypes,DsAcc-method}
\alias{getRegionTypes}
\title{getRegionTypes-methods}
\usage{
\S4method{getRegionTypes}{DsAcc}(.object, inclSites = FALSE)
}
\arguments{
\item{.object}{\code{\linkS4class{DsAcc}} object}

\item{inclSites}{include \code{"sites"} in the result}
}
\value{
Character vector of sample IDs in the dataset
}
\description{
Return sample IDs in a dataset
}
\author{
Fabian Mueller
}
