% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsAcc-class.R
\docType{methods}
\name{getSampleAnnot,DsAcc-method}
\alias{getSampleAnnot,DsAcc-method}
\alias{getSampleAnnot}
\title{getSampleAnnot-methods}
\usage{
\S4method{getSampleAnnot}{DsAcc}(.object)
}
\arguments{
\item{.object}{\code{\linkS4class{DsAcc}} object}
}
\value{
\code{data.frame} containing sample annotation
}
\description{
Return sample annotation table of a dataset
}
\author{
Fabian Mueller
}
