% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsNOMe-class.R
\docType{methods}
\name{getSampleAnnot,DsNOMe-method}
\alias{getSampleAnnot}
\alias{getSampleAnnot,DsNOMe-method}
\title{getSampleAnnot-methods}
\usage{
\S4method{getSampleAnnot}{DsNOMe}(.object)
}
\arguments{
\item{.object}{\code{\linkS4class{DsNOMe}} object}
}
\value{
\code{data.frame} containing sample annotation
}
\description{
Return sample annotation table of a dataset
}
\author{
Fabian Mueller
}

