% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_atac.R
\name{getSampleMetrics.snakeATAC}
\alias{getSampleMetrics.snakeATAC}
\title{getSampleMetrics.snakeATAC}
\usage{
getSampleMetrics.snakeATAC(sampleAnnot, snakeDir, withPeaks = TRUE)
}
\arguments{
\item{sampleAnnot}{data.frame specifying the sample annotation table. Must have valid rownames corresponding to the sample ids used in
the snakeAtac filenames}

\item{snakeDir}{snakeATAC base directory (where the files are located)}

\item{withPeaks}{flag indicating whether to output peak statistics}
}
\value{
\code{data.frame} containing sample summary statistics. the original sample annotation table will be appended to the summary output
}
\description{
Retrieve sample summary statistics from the output a snakeATAC pipline run
}
\author{
Fabian Mueller
}
