% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsAcc-class.R
\docType{methods}
\name{getSamples,DsAcc-method}
\alias{getSamples,DsAcc-method}
\alias{getSamples}
\title{getSamples-methods}
\usage{
\S4method{getSamples}{DsAcc}(.object)
}
\arguments{
\item{.object}{\code{\linkS4class{DsAcc}} object}
}
\value{
Character vector of sample IDs in the dataset
}
\description{
Return sample IDs in a dataset
}
\author{
Fabian Mueller
}
