% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsNOMe-class.R
\docType{methods}
\name{getSamples,DsNOMe-method}
\alias{getSamples,DsNOMe-method}
\alias{getSamples}
\title{getSamples-methods}
\usage{
\S4method{getSamples}{DsNOMe}(.object)
}
\arguments{
\item{.object}{\code{\linkS4class{DsNOMe}} object}
}
\value{
Character vector of sample IDs in the dataset
}
\description{
Return sample IDs in a dataset
}
\author{
Fabian Mueller
}
