% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATACsc-class.R
\docType{methods}
\name{getScQcStatsTab,DsATACsc-method}
\alias{getScQcStatsTab,DsATACsc-method}
\alias{getScQcStatsTab}
\title{getScQcStatsTab-methods}
\usage{
\S4method{getScQcStatsTab}{DsATACsc}(.object)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATACsc}} object}
}
\value{
an \code{data.frame} contain QC statistics for each cell
}
\description{
Retrieve a table of QC statistics for single cells
}
\author{
Fabian Mueller
}
