% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{getTfAnnot}
\alias{getTfAnnot}
\title{getTfAnnot}
\usage{
getTfAnnot(type = "humantfs")
}
\arguments{
\item{type}{annotation type. Currently only \code{"humantfs"} (pulls info from humantfs.ccbr.utoronto.ca) is supported}
}
\value{
a data frame of TF annotation
}
\description{
retrieve TF annotation data
}
\author{
Fabian Mueller
}
