% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{getTssEnrichment,DsATAC-method}
\alias{getTssEnrichment,DsATAC-method}
\alias{getTssEnrichment}
\title{getTssEnrichment-methods}
\usage{
\S4method{getTssEnrichment}{DsATAC}(.object, sampleId, tssGr,
  flank = 2000L, normTailW = 100L, smoothW = 25L)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{sampleId}{sample to be plotted}

\item{tssGr}{\code{GRanges} object containing TSS coordinates}

\item{flank}{number of bases flanking each TSS that will be added on each side}

\item{normTailW}{number of bases on each side whose counts will be used to normalize the data}

\item{smoothW}{radius of the window (in bp) that will be used to smooth the data, i.e. the total width of the
smoothing window will be twice that number}
}
\value{
a list containing TSS enrichment data and a \code{ggplot} object containing TSS enrichment plot
}
\description{
Get TSS enrichment data and plot
}
\author{
Fabian Mueller
}
