% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{getTssEnrichmentBatch,DsATAC-method}
\alias{getTssEnrichmentBatch,DsATAC-method}
\alias{getTssEnrichmentBatch}
\title{getTssEnrichmentBatch-methods}
\usage{
\S4method{getTssEnrichmentBatch}{DsATAC}(.object, tssGr = NULL,
  sampleIds = getSamples(.object), tssW = 201L, flank = 2000L,
  normTailW = 200L, smoothW = 51L)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{tssGr}{\code{GRanges} object containing TSS coordinates or NULL to get default set from annotation package}

\item{sampleIds}{sampleIds for which TSS enrichment should be computed}

\item{flank}{number of bases flanking each TSS that will be added on each side}

\item{normTailW}{number of bases on each side whose counts will be used to normalize the data}

\item{smoothW}{diameter of the window (in bp) that will be used to smooth the data}
}
\value{
a list containing TSS enrichment data
}
\description{
Get TSS enrichment data and plot
}
\author{
Fabian Mueller
}
