% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isCanonicalChrom}
\alias{isCanonicalChrom}
\title{isCanonicalChrom}
\usage{
isCanonicalChrom(ss)
}
\arguments{
\item{ss}{character vector of chromosome names}
}
\value{
logical vector stating whether the given chromosome names correspond to canonical chromosomes
}
\description{
for a character string of chromosome names, determine if it is a canonical chromosome
(i.e. not not ChrUn*, *_random, ...)
}
\author{
Fabian Mueller
}
