% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATACsc-class.R
\docType{methods}
\name{iterativeLSI,DsATACsc-method}
\alias{iterativeLSI,DsATACsc-method}
\alias{iterativeLSI}
\title{iterativeLSI-methods}
\usage{
\S4method{iterativeLSI}{DsATACsc}(.object, it0regionType = "t5k",
  it0nMostAcc = 20000L, it0pcs = 2:25, it0clusterResolution = 0.8,
  it0nTopPeaksPerCluster = 2e+05, it1pcs = 1:50,
  it1clusterResolution = 0.8, it1mostVarPeaks = 50000L,
  it2pcs = 1:50, it2clusterResolution = 0.8)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATACsc}} object}

\item{it0regionType}{character string specifying the region type to start with}

\item{it0nMostAcc}{the number of the most accessible regions to consider in iteration 0}

\item{it0pcs}{the principal components to consider in iteration 0}

\item{it0clusterResolution}{resolution paramter for Seurat's  clustering (\code{Seurat::FindClusters}) in iteration 0}

\item{it0nTopPeaksPerCluster}{the number of best peaks to be considered for each cluster in the merged peak set (iteration 0)}

\item{it1pcs}{the principal components to consider in iteration 0}

\item{it1clusterResolution}{resolution paramter for Seurat's  clustering (\code{Seurat::FindClusters}) in iteration 1}

\item{it1mostVarPeaks}{the number of the most variable peaks to consider after iteration 1}

\item{it2pcs}{the principal components to consider in the final iteration (2)}

\item{it2clusterResolution}{resolution paramter for Seurat's  clustering (\code{Seurat::FindClusters}) in the final iteration (2)}
}
\value{
an \code{S3} object containing dimensionality reduction results and clustering
}
\description{
EXPERIMENTAL: Perform iterative LSI clustering as described in doi:10.1101/696328
}
\author{
Fabian Mueller
}
