% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{join,DsATAC-method}
\alias{join,DsATAC-method}
\alias{join}
\title{join-methods}
\usage{
\S4method{join}{DsATAC}(.object, objectB, joinRegionTypes = "union")
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{objectB}{\code{\linkS4class{DsATAC}} object}

\item{combineRegionTypes}{how to combine region types:
'union' (default): the resulting object will have counts aggregated over region types
from both objects. 'intersect': only region types present in both objects will occur
   in the output}
}
\value{
a new \code{\linkS4class{DsATAC}} object combining both input objects.
        It contains untransformed counts.
}
\description{
Combine two \code{\linkS4class{DsATAC}} objects
}
\author{
Fabian Mueller
}
