% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsAcc-class.R
\name{length,DsAcc-method}
\alias{length,DsAcc-method}
\title{Retrieve the number of samples contained in a DsAcc object}
\usage{
\S4method{length}{DsAcc}(x)
}
\arguments{
\item{x}{DsAcc object}
}
\description{
Retrieve the number of samples contained in a DsAcc object
}
