% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{loadConfig}
\alias{loadConfig}
\title{loadConfig}
\usage{
loadConfig(cfgFile)
}
\arguments{
\item{cfgFile}{Config file in JSON format. As output by \link{saveConfig}}
}
\value{
nothing of particular interest. The configuration is set for the current environment
}
\description{
Sets the configuration from a configuration file (JSON)
}
\author{
Fabian Mueller
}
