% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsAcc-class.R
\name{loadDsAcc}
\alias{loadDsAcc}
\title{loadDsAcc}
\usage{
loadDsAcc(path)
}
\arguments{
\item{path}{Location of saved \code{\linkS4class{DsAcc}} object}
}
\value{
\code{\linkS4class{DsAcc}} object
}
\description{
Load a DsAcc dataset from disk
}
\author{
Fabian Mueller
}
