% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsNOMe-class.R
\name{loadDsNOMe}
\alias{loadDsNOMe}
\title{loadDsNOMe}
\usage{
loadDsNOMe(path)
}
\arguments{
\item{path}{Location of saved \code{\linkS4class{DsNOMe}} object}
}
\value{
\code{\linkS4class{DsNOMe}} object
}
\description{
Load a DsNOMe dataset from disk
}
\author{
Fabian Mueller
}

