% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsNOMe-class.R
\docType{methods}
\name{maskMethNA,DsNOMe-method}
\alias{maskMethNA}
\alias{maskMethNA,DsNOMe-method}
\title{maskMethNA-methods}
\usage{
\S4method{maskMethNA}{DsNOMe}(.object, mask, type = "sites",
  reaggregate = TRUE)
}
\arguments{
\item{.object}{\code{\linkS4class{DsNOMe}} object}

\item{mask}{a mask, i.e. a logical matrix of indices to set to NA}

\item{type}{character string specifying a name for the region type (sefault: sites)}

\item{reaggregate}{redo region aggregation (only has an effect if type is sites and there are aggregated regions in the dataset)}
}
\value{
a new \code{\linkS4class{DsNOMe}} object with sites/regions masked
}
\description{
Set the indices specified in a mask to NA
}
\author{
Fabian Mueller
}

