% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{mergeSamples,DsATAC-method}
\alias{mergeSamples,DsATAC-method}
\alias{mergeSamples}
\title{mergeSamples-methods}
\usage{
\S4method{mergeSamples}{DsATAC}(.object, mergeGroups,
  countAggrFun = "median")
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{mergeGroups}{factor or character vector or column name in sample annotation table}

\item{countAggrFun}{aggregation function for signal counts.
Currently \code{sum}, \code{mean} and \code{median} (default) are supported.}
}
\value{
a new \code{\linkS4class{DsATAC}} object with samples merged
}
\description{
Merge signal and insertion data across samples
}
\author{
Fabian Mueller
}
