% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{normalizeCounts,DsATAC-method}
\alias{normalizeCounts,DsATAC-method}
\alias{normalizeCounts}
\title{normalizeCounts-methods}
\usage{
\S4method{normalizeCounts}{DsATAC}(.object, method = "quantile",
  regionTypes = getRegionTypes(.object))
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{method}{normalization method to be applied. Currently only 'quantile' is supported}

\item{regionTypes}{character vector specifying a name for the region type in which count data should be normalized(default: all region types)}
}
\value{
a new \code{\linkS4class{DsATAC}} object with normalized count data
}
\description{
Normalize methylation levels
}
\author{
Fabian Mueller
}
