% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsNOMe-class.R
\docType{methods}
\name{normalizeMeth,DsNOMe-method}
\alias{normalizeMeth,DsNOMe-method}
\alias{normalizeMeth}
\title{normalizeMeth-methods}
\usage{
\S4method{normalizeMeth}{DsNOMe}(.object, type = "sites",
  method = "quantile")
}
\arguments{
\item{.object}{\code{\linkS4class{DsNOMe}} object}

\item{type}{character string specifying a name for the region type (default: sites)}

\item{method}{normalization method to be applied. Currently only 'quantile' is supported}

\item{reaggregate}{redo region aggregation (only has an effect if type is sites and there are aggregated regions in the dataset)}
}
\value{
a new \code{\linkS4class{DsNOMe}} object with normalized methylation levels
}
\description{
Normalize methylation levels
}
\author{
Fabian Mueller
}
