% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsASC-tfactivity.R
\name{plotASCTFHandoff}
\alias{plotASCTFHandoff}
\title{Handoff plot: change in allele-specific TF activity between two contexts}
\usage{
plotASCTFHandoff(switchDt, topN = 15L)
}
\arguments{
\item{switchDt}{output of \code{ascTFActivitySwitch}.}

\item{topN}{number of TFs (by |deltaObs|) to label (default 15).}
}
\value{
a ggplot.
}
\description{
Handoff plot: change in allele-specific TF activity between two contexts
}
\author{
Irem B. GUNDUZ
}
