% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{plotInsertSizeDistribution,DsATAC-method}
\alias{plotInsertSizeDistribution,DsATAC-method}
\alias{plotInsertSizeDistribution}
\title{plotInsertSizeDistribution-methods}
\usage{
\S4method{plotInsertSizeDistribution}{DsATAC}(.object, sampleId)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{sampleId}{sample to be plotted}
}
\value{
\code{ggplot} object containing insert size distribution plot
}
\description{
Plot insert size distribution
}
\author{
Fabian Mueller
}
