% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{projectMatrix_UMAP}
\alias{projectMatrix_UMAP}
\title{projectMatrix_UMAP}
\usage{
projectMatrix_UMAP(X, umapObj, binarize = TRUE)
}
\arguments{
\item{X}{matrix to be projected (features X samples)}

\item{umapObj}{dimension reduction result as returned by \code{\link{dimRed_UMAP}}}

\item{binarize}{binarize the counts before projecting}
}
\value{
Projected UMAP coordinates
}
\description{
given a (count) matrix and dimension reduction result, return the projected UMAP coordinates
in the embedding space
}
\author{
Fabian Mueller
}
