% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_motifs.R
\name{PWMatrixToProbMatrix}
\alias{PWMatrixToProbMatrix}
\title{PWMatrixToProbMatrix}
\usage{
PWMatrixToProbMatrix(x)
}
\arguments{
\item{x}{log2probratio PWM (\code{PWMatrix} from TFBSTools package)}
}
\value{
PWM probability matrix with values in
}
\description{
convert a log2probratio PWM (\code{PWMatrix} from TFBSTools package) to a matrix containing probabilities in [0,1]
}
\author{
Fabian Mueller
}
