% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_atac.R
\name{readMACS2peakFile}
\alias{readMACS2peakFile}
\title{readMACS2peakFile}
\usage{
readMACS2peakFile(fn)
}
\arguments{
\item{fn}{Filename for MACS2 narrow peak file}
}
\value{
\code{GRanges} object containing peak information
}
\description{
Reads the MACS2 ouput as GRanges
}
\author{
Fabian Mueller
}
