% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{regionSetCounts,DsATAC-method}
\alias{regionSetCounts,DsATAC-method}
\alias{regionSetCounts}
\title{regionSetCounts-methods}
\usage{
\S4method{regionSetCounts}{DsATAC}(.object, rsl, bySample = FALSE)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{rsl}{\code{GRangesList} or NAMED list of \code{GRanges} objects. Each element corresponds to a region set for which the summary statistics are reported}

\item{bySample}{for internal use: iterate over samples (instead of retrieving one giant insertion list for all samples) in order to save memory (at the tradeoff of compute time)}
}
\value{
a matrix of overlap counts for each region set and sample
}
\description{
Overlap the insertion data with a list of region sets
}
\author{
Fabian Mueller
}
