% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{removeFragmentData,DsATAC-method}
\alias{removeFragmentData,DsATAC-method}
\alias{removeFragmentData}
\title{removeFragmentData-methods}
\usage{
\S4method{removeFragmentData}{DsATAC}(object)
}
\arguments{
\item{object}{\code{\linkS4class{DsATAC}} object}
}
\value{
the modified object (without fragment data)
}
\description{
Removes fragment data from \code{\linkS4class{DsATAC}} object (e.g. to save space)
}
\author{
Fabian Mueller
}
