% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{removeRegionData,DsATAC-method}
\alias{removeRegionData,DsATAC-method}
\alias{removeRegionData}
\title{removeRegionData-methods}
\usage{
\S4method{removeRegionData}{DsATAC}(.object)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}
}
\value{
a new \code{\linkS4class{DsATAC}} object with region data removed
}
\description{
Remove all region data from a \code{\linkS4class{DsATAC}} object
}
\author{
Fabian Mueller
}
