% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsNOMe-class.R
\docType{methods}
\name{removeRegions,DsNOMe-method}
\alias{removeRegions,DsNOMe-method}
\title{removeRegions-methods}
\usage{
\S4method{removeRegions}{DsNOMe}(.object, indices, type = "sites",
  reaggregate = TRUE)
}
\arguments{
\item{.object}{\code{\linkS4class{DsNOMe}} object}

\item{indices}{a vector of indices of sites/regions to be removed. Can be numeric, integer or logical.}

\item{type}{character string specifying a name for the region type (sefault: sites)}

\item{reaggregate}{redo region aggregation (only has an effect if type is sites and there are aggregated regions in the dataset)}
}
\value{
a new \code{\linkS4class{DsNOMe}} object with sites/regions removed
}
\description{
Remove the specified sites or regions from an object
}
\author{
Fabian Mueller
}
