% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{removeRegionType,DsATAC-method}
\alias{removeRegionType,DsATAC-method}
\alias{removeRegionType}
\title{removeRegionType-methods}
\usage{
\S4method{removeRegionType}{DsATAC}(.object, type)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{type}{character string specifying a name for the region type (sefault: sites)}
}
\value{
a new \code{\linkS4class{DsATAC}} object with the region type removed
}
\description{
Remove the specified region type from an object
}
\author{
Fabian Mueller
}
