% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{removeSamples,DsATAC-method}
\alias{removeSamples,DsATAC-method}
\alias{removeSamples}
\title{removeSamples-methods}
\usage{
\S4method{removeSamples}{DsATAC}(.object, indices)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{indices}{a vector of indices of samples to be removed. Can be numeric, integer or logical.}
}
\value{
a new \code{\linkS4class{DsATAC}} object with sites/regions removed
}
\description{
Remove samples from a \code{\linkS4class{DsATAC}} object
}
\author{
Fabian Mueller
}
