% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowZscores}
\alias{rowZscores}
\title{rowZscores}
\usage{
rowZscores(X, na.rm = FALSE)
}
\arguments{
\item{X}{input matrix}
}
\value{
z-score normalized matrix
}
\description{
Performs z-score normalization on the rows of a matrix. (Basically a wrapper around \code{matrixStats})
}
\author{
Fabian Mueller
}
