% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_atac_chromvar}
\alias{run_atac_chromvar}
\title{run_atac_chromvar}
\usage{
run_atac_chromvar(.object)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}
}
\value{
An S3 object containing a list of chromVAR Deviations objects as returned by \code{chromVAR::computeDeviations}. One object for each region type specified in the \code{chromVarRegionTypes} configuration.
}
\description{
Run chromVAR analysis for ATAC-seq data
}
\author{
Fabian Mueller
}
