% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_atac_exploratory}
\alias{run_atac_exploratory}
\title{run_atac_exploratory}
\usage{
run_atac_exploratory(dsa, anaDir, itLsiObj = NULL)
}
\arguments{
\item{dsa}{\code{\linkS4class{DsATAC}} object}

\item{anaDir}{analysis directory}

\item{itLsiObj}{[for single-cell only; optional] pre-computed result of a call to \code{iterativeLSI(.object, ...)}}
}
\value{
\code{S3} object containing exploratory metrics and an analysis report object
}
\description{
Run exploratory analyses for ATAC-seq data
}
\author{
Fabian Mueller
}
