% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_atac_filtering}
\alias{run_atac_filtering}
\title{run_atac_filtering}
\usage{
run_atac_filtering(dsa, anaDir)
}
\arguments{
\item{dsa}{\code{\linkS4class{DsATAC}} object}

\item{anaDir}{analysis directory}
}
\value{
\code{S3} object containing the filtered \code{\linkS4class{DsATAC}} object, filtering statistics and an analysis report object
}
\description{
Run the filtering for ATAC-seq data
}
\author{
Fabian Mueller
}
