% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_atac_normalization}
\alias{run_atac_normalization}
\title{run_atac_normalization}
\usage{
run_atac_normalization(dsa, anaDir)
}
\arguments{
\item{dsa}{\code{\linkS4class{DsATAC}} object}

\item{anaDir}{analysis directory}
}
\value{
\code{S3} object containing the normalized \code{\linkS4class{DsATAC}} object and an analysis report object
}
\description{
Run count normalization for ATAC-seq data
}
\author{
Fabian Mueller
}
