% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_atac_peakcalling}
\alias{run_atac_peakcalling}
\title{run_atac_peakcalling}
\usage{
run_atac_peakcalling(dsa, anaDir)
}
\arguments{
\item{dsa}{\code{\linkS4class{DsATAC}} object}

\item{anaDir}{analysis directory}
}
\value{
\code{S3} object containing the annotated \code{\linkS4class{DsATAC}} object, per-sample peak calls, a consensus peak set and an analysis report object
}
\description{
Run peak calling for ATAC-seq data
}
\author{
Fabian Mueller
}
