% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_atac_qc}
\alias{run_atac_qc}
\title{run_atac_qc}
\usage{
run_atac_qc(dsa, anaDir)
}
\arguments{
\item{dsa}{\code{\linkS4class{DsATAC}} object}

\item{anaDir}{analysis directory}
}
\value{
\code{S3} object containing QC statistics and an analysis report object
}
\description{
Run the summary QC analysis for ATAC-seq data
}
\author{
Fabian Mueller
}
