% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_atac_sc_unsupervised}
\alias{run_atac_sc_unsupervised}
\title{run_atac_sc_unsupervised}
\usage{
run_atac_sc_unsupervised(dsa, anaDir)
}
\arguments{
\item{dsa}{\code{\linkS4class{DsATAC}} object}

\item{anaDir}{analysis directory}
}
\value{
\code{S3} object containing the annoted \code{\linkS4class{DsATAC}} object, the results of running iterative LSI and an analysis report object
}
\description{
Run unsupervised analysis for single-cell ATAC-seq data (i.e. iterative LSI, clustering and cluster peak detection)
}
\author{
Fabian Mueller
}
