% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safeMatrixStats}
\alias{safeMatrixStats}
\title{safeMatrixStats}
\usage{
safeMatrixStats(X, statFun = "rowSums", ...)
}
\arguments{
\item{X}{input matrix}

\item{statFun}{statistic. E.g. \code{"rowSums"}, \code{"colSums"}, \code{"rowMeans"}, \code{"colMeans"}, ...}

\item{...}{arguments passed on to the matrix stats function. E.g. \code{na.rm}.}
}
\value{
result of the corresponding matrix statistic
}
\description{
Compute matrix statistics selecting the appropriate function depending on the
matrix class of the input (supports sparse matrices and DelayedArrays)
}
\author{
Fabian Mueller
}
