% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATACsc-class.R
\docType{methods}
\name{samplePseudoBulk,DsATACsc-method}
\alias{samplePseudoBulk,DsATACsc-method}
\alias{samplePseudoBulk}
\title{samplePseudoBulk-methods}
\usage{
\S4method{samplePseudoBulk}{DsATACsc}(.object, nnData, nSamples, nCellsPerSample = 100)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATACsc}} object}

\item{nnData}{Data to use for nearest neighbor matching. Can either be the
name of a region type in \code{.object} or a data matrix with
the same number of rows as \code{.object} has cells.}

\item{nSamples}{number of pseudobulk samples to be returned}

\item{nCellsPerSample}{number of cells to be aggregated per sample}
}
\value{
\code{S3} data structure containing a list of sampling results as well as
        a \code{\linkS4class{DsATAC}} object containing pseudo-bulk aggregates
}
\description{
Samples pseudo-bulk samples from single-cells
}
\details{
Samples pseudo-bulk samples from single-cells by sampling \code{nSamples} individual cells
and then merging it with its \code{nCellsPerSample - 1} nearest neighbors (according to \code{nnData}).
}
\author{
Fabian Mueller
}
