% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{saveConfig}
\alias{saveConfig}
\title{saveConfig}
\usage{
saveConfig(dest)
}
\arguments{
\item{dest}{Filename for the config file in JSON format.}
}
\value{
nothing of particular interest.
}
\description{
Save the current configuration to a configuration file (JSON)
}
\author{
Fabian Mueller
}

