% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsNOMe-class.R
\name{saveDsNOMe}
\alias{saveDsNOMe}
\title{saveDsNOMe}
\usage{
saveDsNOMe(.object, path)
}
\arguments{
\item{.object}{\code{\linkS4class{DsNOMe}} object}

\item{path}{destination to save the object to}
}
\value{
nothing of particular interest
}
\description{
Save a DsNOMe dataset to disk for later loading
}
\author{
Fabian Mueller
}
