% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsASC-class.R
\name{setAccessibility}
\alias{setAccessibility}
\title{Attach a peak accessibility matrix (+ peak coordinates) to a DsASC object}
\usage{
setAccessibility(dsObj, peakMat, peakGr)
}
\arguments{
\item{dsObj}{A \code{\linkS4class{DsASC}} object.}

\item{peakMat}{Matrix of peak counts [peak x sample]. rownames = peak IDs,
colnames must contain every \code{getSamples(dsObj)}.}

\item{peakGr}{GRanges of consensus peaks. names(peakGr) must match
rownames(peakMat).}
}
\value{
The updated DsASC object.
}
\description{
Populates the \code{accessibility} slot and \code{coord$peaks}. The peak
matrix is typically produced by ChrAccR / your 02_ChrAccR.R step.
}
\author{
Irem B. GUNDUZ
}
