% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATACsc-class.R
\docType{methods}
\name{simulateDoublets,DsATAC-method}
\alias{simulateDoublets,DsATAC-method}
\alias{simulateDoublets}
\title{simulateDoublets-methods}
\usage{
\S4method{simulateDoublets}{DsATAC}(.object, byGroup = NULL, n = 10000,
  sampleRatio = 1)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATACsc}} object}

\item{byGroup}{sample by group. Must be a column name in the sample annotation table}

\item{n}{number of doublets to simulate (per group)}

\item{sampleRatio}{fraction of non-zero events to subsample. If \code{0 < sampleRatio < 1}, the individual cell counts will be subsampled
for each region.}
}
\value{
an \code{\linkS4class{DsATACsc}} object containing counts for simulated doublets. Fragment data will be discarded.
}
\description{
EXPERIMENTAL: Simulate doublets by adding counts in matrices for each region set
}
\author{
Fabian Mueller
}
