% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smoothMagic}
\alias{smoothMagic}
\title{smoothMagic}
\usage{
smoothMagic(X, X_knn = NULL, k = 15, ka = ceiling(k/4), td = 3)
}
\arguments{
\item{X}{matrix to be smoothed (data points [cells] represent rows and features [genes] represent columns)}

\item{X_knn}{matrix to be used for finding nearest neighbors. Must have the same rows (cells) as X}

\item{k}{number of nearest neighbors to compute}

\item{ka}{The adaptive kernel width (sigma) will be set to the distance to the \code{ka}th nearest neighbor for each cell. Thus the kernel is wider in sparse areas and smaller in dense areas.}

\item{td}{diffusion time (number of exponentiations of the kernel matrix)}
}
\value{
list containing the smoothed matrix and the kernel matrix
}
\description{
[!EXPERIMENTAL!] smooth a matrix using an adaptation of the MAGIC algorithm (doi:10.1016/j.cell.2018.05.061)
}
\author{
Jeff Granja, Fabian Mueller
}
