% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\name{[,DsATAC,ANY,ANY,ANY-method}
\alias{[,DsATAC,ANY,ANY,ANY-method}
\title{Subsetting DsATAC datasets by sample}
\usage{
\S4method{[}{DsATAC,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{DsATAC object}

\item{i}{sample names or indices}
}
\description{
NOTE: '[' operator for DsATAC does not reorder samples or deal with index multiplicity
}
