% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{unsupervisedAnalysisSc,DsATAC-method}
\alias{unsupervisedAnalysisSc,DsATAC-method}
\alias{unsupervisedAnalysisSc}
\title{unsupervisedAnalysisSc-methods}
\usage{
\S4method{unsupervisedAnalysisSc}{DsATAC}(.object, regionType,
  regionIdx = NULL, dimRedMethod = "tf-idf_irlba", usePcs = 1:50,
  clusteringMethod = "seurat_louvain")
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{regionType}{character string specifying the region type}

\item{regionIdx}{indices of regions to be used (logical or integer vector). If \code{NULL} (default) all regions of the specified regionType will be used.}

\item{dimRedMethod}{character string specifying the dimensionality reduction method. Currently on \code{"tf-idf_irlba"} is supported}

\item{usePcs}{integer vector specifying the principal components to use for UMAP and clustering}

\item{clusteringMethod}{character string specifying the clustering method. Currently on \code{"seurat_louvain"} is supported}
}
\value{
an \code{S3} object containing dimensionality reduction results and clustering
}
\description{
Perform unsupervised analysis on single-cell data. Performs dimensionality reduction
and clustering.
}
\author{
Fabian Mueller
}
