% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colZscores}
\alias{colZscores}
\title{colZscores}
\usage{
colZscores(X, na.rm = FALSE)
}
\arguments{
\item{X}{input matrix}

\item{na.rm}{should NAs be omitted?}
}
\value{
z-score normalized matrix
}
\description{
Performs z-score normalization on the columns of a matrix. (Basically a wrapper around \code{matrixStats})
}
\author{
Fabian Mueller
}
