% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_atac_exploratory.R
\docType{methods}
\name{createReport_exploratory,DsATAC-method}
\alias{createReport_exploratory,DsATAC-method}
\alias{createReport_exploratory}
\title{createReport_exploratory-methods}
\usage{
\S4method{createReport_exploratory}{DsATAC}(.object, reportDir)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{reportDir}{directory in which the report will be created}
}
\value{
(invisible) \code{muReportR::Report} object containing the report
}
\description{
Create a report summarizing exploratory analyses of an accessibility dataset
}
\author{
Fabian Mueller
}
