% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsAcc-class.R
\docType{class}
\name{DsAcc-class}
\alias{DsAcc-class}
\title{DsAcc}
\description{
A class for accessibility datasets
}
\section{Slots}{

\describe{
  \item{\code{coord}}{
	List of coordinates (GRanges objects) for accessibility summarized regions.
  }
  \item{\code{sampleAnnot}}{
	Sample annotation Table
  }
  \item{\code{genome}}{
	Genome assembly
  }
}
}

\section{Methods}{

\describe{
   \item{\code{\link{samples,DsAcc-method}}}{
     Retrieve a vector of sample IDs for the dataset
   }
}
}

\author{
Fabian Mueller
}
