% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsAcc-class.R
\docType{class}
\name{DsAcc-class}
\alias{DsAcc-class}
\title{DsAcc}
\description{
A class for accessibility datasets
}
\section{Slots}{

\describe{
  \item{\code{coord}}{
	List of coordinates (GRanges objects) for accessibility summarized regions.
  }
  \item{\code{sampleAnnot}}{
	Sample annotation Table
  }
  \item{\code{genome}}{
	Genome assembly
  }
  \item{\code{diskDump}}{
	Flag indicating whether large matrices and objects will be kept on disk rather than in main memory.
  }
}
}

\author{
Fabian Mueller
}
