% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{class}
\name{DsATAC-class}
\alias{DsATAC-class}
\title{DsATAC}
\description{
A class for storing NOMe accessibility data
}
\section{Slots}{

\describe{
  \item{\code{covg}}{
	List of GC read coverage for sites and summarized regions.
  }
}
}

\section{Methods}{

\describe{
   \item{\code{\link{samples,DsATAC-method}}}{
     Retrieve a vector of sample IDs for the dataset
   }
}
}

\author{
Fabian Mueller
}
