% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsNOMe-class.R
\docType{class}
\name{DsNOMe-class}
\alias{DsNOMe-class}
\title{DsNOMe}
\description{
A class for storing information on pipeline jobs.
}
\section{Slots}{

\describe{
  \item{\code{coord}}{
	List of coordinates (GRanges objects) for GC dinucleotides (sites) and summarized regions.
  }
  \item{\code{meth}}{
	List of GC methylation for sites and summarized regions.
  }
  \item{\code{covg}}{
	List of GC read coverage for sites and summarized regions.
  }
  \item{\code{sampleAnnot}}{
	Sample annotation Table
  }
  \item{\code{genome}}{
	Genome assembly
  }
}
}

\section{Methods}{

\describe{
   \item{\code{\link{samples,DsNOMe-method}}}{
     Retrieve a vector of sample IDs for the dataset
   }
}
}

\author{
Fabian Mueller
}
