% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsASC-class.R
\name{filterASCByPeaks}
\alias{filterASCByPeaks}
\title{Filter DsASC to Peak-Overlapping SNPs}
\usage{
filterASCByPeaks(dsObj)
}
\arguments{
\item{dsObj}{A DsASC object (must be annotated first).}
}
\value{
A subsetted DsASC object.
}
\description{
Removes SNPs that do not overlap with any consensus peak.
This reduces dataset size and focuses analysis on regulatory regions.
}
