% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsASC-class.R
\name{filterLowCovg,DsASC-method}
\alias{filterLowCovg,DsASC-method}
\title{filterLowCovg-methods (DsASC)}
\usage{
\S4method{filterLowCovg}{DsASC}(.object, thresh = 1L, reqSamples = 0.75)
}
\arguments{
\item{.object}{A \code{\linkS4class{DsASC}} object.}

\item{thresh}{Integer. Minimum total coverage required per site per sample.}

\item{reqSamples}{Numeric. Minimum number (>=1) or fraction (<1) of samples
that must meet the coverage threshold.}
}
\description{
filterLowCovg-methods (DsASC)
}
\author{
Irem B. GUNDUZ
}
