% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{getCounts,DsATAC-method}
\alias{getCounts,DsATAC-method}
\alias{getCounts}
\title{getCounts-methods}
\usage{
\S4method{getCounts}{DsATAC}(.object, type, i = NULL, j = NULL,
  asMatrix = TRUE)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{type}{character string specifying the region type}

\item{i}{(optional) row (region) indices}

\item{j}{(optional) column (sample) indices}

\item{asMatrix}{return a matrix object instead of the internal representation}
}
\value{
\code{data.table} or \code{matrix} containing counts for
        each region and sample
}
\description{
Return table of count values
}
\author{
Fabian Mueller
}
