% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{getCounts,DsATAC-method}
\alias{getCounts,DsATAC-method}
\alias{getCounts}
\title{getCounts-methods}
\usage{
\S4method{getCounts}{DsATAC}(
  .object,
  type,
  i = NULL,
  j = NULL,
  asMatrix = TRUE,
  naIsZero = TRUE,
  allowSparseMatrix = FALSE
)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{type}{character string specifying the region type}

\item{i}{(optional) row (region) indices}

\item{j}{(optional) column (sample) indices}

\item{asMatrix}{return a matrix object instead of the internal representation}

\item{naIsZero}{should \code{NA}s in the count matrix be considered 0 value (instead of unknown/missing)}

\item{allowSparseMatrix}{if \code{asMatrix}: allow for sparse matrices as returned data format}
}
\value{
Matrix containing counts for
        each region and sample
}
\description{
Return table of count values
}
\author{
Fabian Mueller
}
