% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chromVAR.R
\name{getGenomeObject}
\alias{getGenomeObject}
\title{getGenomeObject}
\usage{
getGenomeObject(assembly, adjChrNames = TRUE)
}
\arguments{
\item{assembly}{string specifying the assembly}

\item{adjChrNames}{should the prefix "chr" be added to main chromosomes if not already present and chrMT be renamed to chrM?}
}
\value{
\code{BSgenome} object
}
\description{
retrieve the appropriate \code{BSgenome} for an assembly string
}
