% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{getInsertionSites,DsATAC-method}
\alias{getInsertionSites,DsATAC-method}
\alias{getInsertionSites}
\title{getInsertionSites-methods}
\usage{
\S4method{getInsertionSites}{DsATAC}(.object,
  samples = getSamples(.object), asGRangesList = FALSE)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{samples}{sample identifiers}

\item{asGRangesList}{should a \code{GRangesList} object be returned instead of a regular \code{list}}
}
\value{
\code{list} or \code{GRangesList} containing Tn5 cut sites for each sample
}
\description{
Return a list of insertion sites (Tn5 cut sites) for each sample
}
\author{
Fabian Mueller
}
