% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsNOMe-class.R
\docType{methods}
\name{mergeStrands,DsNOMe-method}
\alias{mergeStrands}
\alias{mergeStrands,DsNOMe-method}
\title{mergeStrands-methods}
\usage{
\S4method{mergeStrands}{DsNOMe}(.object)
}
\arguments{
\item{.object}{\code{\linkS4class{DsNOMe}} object}
}
\value{
a new \code{\linkS4class{DsNOMe}} object with the strands merged
}
\description{
Merge + and - strands of the dataset by adding read coverage and recomputing
Methylation levels
}
\author{
Fabian Mueller
}

