% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsATAC-class.R
\docType{methods}
\name{regionAggregation,DsATAC-method}
\alias{regionAggregation,DsATAC-method}
\alias{regionAggregation}
\title{regionAggregation-methods}
\usage{
\S4method{regionAggregation}{DsATAC}(.object, regGr, type, signal = NULL,
  aggrFun = "median", dropEmpty = TRUE)
}
\arguments{
\item{.object}{\code{\linkS4class{DsATAC}} object}

\item{regGr}{\code{GRanges} object containing regions to summarize}

\item{type}{character string specifying a name for the region type}

\item{signal}{character string specifying a name for the region type for the signal to be aggregated
if it is \code{NULL} (default), the new region type will be initialized with NA values}

\item{aggrFun}{aggregation function for signal counts.
Currently \code{sum}, \code{mean} and \code{median} (default) are supported.}

\item{dropEmpty}{discard all regions with no observed signal counts}
}
\value{
a new \code{\linkS4class{DsATAC}} object with aggregated signal counts per regions
}
\description{
Aggregate signal counts across a set of regions
}
\author{
Fabian Mueller
}
