% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsAcc-class.R
\name{saveDsAcc}
\alias{saveDsAcc}
\title{saveDsAcc}
\usage{
saveDsAcc(.object, path)
}
\arguments{
\item{.object}{\code{\linkS4class{DsAcc}} object}

\item{path}{destination to save the object to}
}
\value{
nothing of particular interest
}
\description{
Save a DsAcc dataset to disk for later loading
}
\author{
Fabian Mueller
}
