% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsAcc-class.R
\name{saveDsAcc}
\alias{saveDsAcc}
\title{saveDsAcc}
\usage{
saveDsAcc(.object, path, forceDiskDump = FALSE, updateDiskRef = TRUE)
}
\arguments{
\item{.object}{\code{\linkS4class{DsAcc}} object}

\item{path}{destination to save the object to}

\item{forceDiskDump}{force large matrices (counts) to be stored as HDF5 (even when the object was not created using \code{diskDump=TRUE})}

\item{updateDiskRef}{update disk dumped (HDF5) references (e.g. for count data)}
}
\value{
(invisibly) The object (with potentially updated disk dumped references)
}
\description{
Save a DsAcc dataset to disk for later loading
}
\author{
Fabian Mueller
}
