% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{setConfigElement}
\alias{setConfigElement}
\title{setConfigElement}
\usage{
setConfigElement(name, value)
}
\arguments{
\item{name}{name of the config item}

\item{value}{value of the config item}
}
\value{
nothing of particular interest.
}
\description{
Set a configuration item to a given value
}
\section{Options used by the package}{

\describe{
  \item{\bold{\code{tmpDir}}\code{ = temdir()}}{
       \code{Directory for temporary files. Must be existing.}
  }
  \item{\bold{\code{cleanMem}}\code{ = TRUE}}{
       \code{During runtime, regularly clean-out the memory in order to reduce memory overuse}
  }
}
}

\author{
Fabian Mueller
}
